feat: set chrome as default platform
feat: change port from setting
